+++ CHANGELOG.md
@@ -1,6 +1,41 @@
 Changelog
 =========
 
+## 3.0.0
+
+### Features
+
+- Added hardware wallet support for "Ledger Nano S", "Ledger Nano X", and "Trezor model T" devices ([PR 2046](https://github.com/input-output-hk/daedalus/pull/2046))
+- Added hardware wallet support for Testnet network ([PR 2264](https://github.com/input-output-hk/daedalus/pull/2264))
+
+### Chores
+
+- Reduced the size of Linux installer ([PR 2260](https://github.com/input-output-hk/daedalus/pull/2260))
+
+## 3.0.0-FC4
+
+### Fixes
+
+- Fixed Trezor transactions TTL ([PR 2046](https://github.com/input-output-hk/daedalus/pull/2046))
+
+## 3.0.0-FC3
+
+### Features
+
+- Added hardware wallet support for "Ledger Nano S" and "Ledger Nano X" on Linux ([PR 2046](https://github.com/input-output-hk/daedalus/pull/2046))
+
+## 3.0.0-FC2
+
+### Features
+
+- Added hardware wallet support for "Ledger Nano S" and "Ledger Nano X" on macOS and Windows ([PR 2046](https://github.com/input-output-hk/daedalus/pull/2046))
+
+## 3.0.0-FC1
+
+### Features
+
+- Added hardware wallet support for "Trezor model T" ([PR 2046](https://github.com/input-output-hk/daedalus/pull/2046))
+
 ## 2.6.0
 
 ### Features